apps/nano_agent_mcp_server/pyproject.toml
[project]
name = "nano-agent"
version = "0.1.0"
description = "MCP server for autonomous engineering agents"
readme = "../../README.md"
authors = [
    { name = "IndyDevDan", email = "agentic@indydevdan.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.4",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0", # Note: >=1.99.2 has Union type issue with openai-agents
    "openai-agents>=0.1.0", # Has compatibility issue with openai>=1.99.2, fixed via typing_fix.py
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.6.7",
    "langchain-openai>=0.3.33",
    "langchain-anthropic>=0.3.20",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
nano-agent = "nano_agent.__main__:run"
nano-cli = "nano_agent.cli:main"

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
